remove cadmodel, use buerli.io/cc/classcad
