+++ client/src/components/Main.tsx
@@ -1,11 +1,12 @@
 import { api as buerliApi, ObjectID } from '@buerli.io/core'
 import { BuerliGeometry, useBuerli } from '@buerli.io/react'
+import { ClassCAD } from '@buerli.io/classcad'
 import { GizmoHelper, GizmoViewcube, GizmoViewport } from '@react-three/drei'
 import { Canvas } from '@react-three/fiber'
 import React from 'react'
 import * as THREE from 'three'
 import { CanvasContainer, ExampleLayout, Spin } from '.'
-import { CadModel } from '../CadModel'
+
 import { storeApi, useStore } from '../store'
 import { Code } from './Code'
 import { Resizer, useResizeStore } from './Resizer'
@@ -98,7 +99,7 @@ const Part: React.FC = () => {
   const params = useStore(s => s.examples.objs[exampleId].params)
   const [meshes, setMeshes] = React.useState<THREE.Mesh[]>([])
   const [scene] = React.useState(() => new THREE.Scene())
-  const model = React.useRef<CadModel>()
+  const model = React.useRef<ClassCAD>()
   const productOrSolidIds = React.useRef<ObjectID | ObjectID[]>(0)
   const fit = useFit(f => f.fit)
   const setModel = useStore(s => s.setModel)
@@ -118,8 +119,8 @@ const Part: React.FC = () => {
     set({ busy: true })
 
     const run = async () => {
-      const m = new CadModel()
-      await m.init()
+      const m = new ClassCAD()
+      await m.connect()
       setModel(exampleId, m)
       model.current = m
       try {
