+++ client/src/models/history/CreatePart.ts
@@ -1,7 +1,7 @@
 import * as THREE from 'three'
 import { Color } from 'three'
 import { Create, GetScene, Param } from '../../store'
-import { setObjectColor, setObjectTransparency } from '../../utils/utils'
+import { setObjectColor, setObjectTransparency } from '../../utils'
 
 export const paramsMap: Param[] = [].sort((a, b) => a.index - b.index)
