+++ client/src/models/solid/polyline1.ts
@@ -2,7 +2,7 @@ import { ObjectID } from '@buerli.io/core'
 import * as THREE from 'three'
 import { Color } from 'three'
 import { Create, GetScene, Param } from '../../store'
-import { setObjectColor } from '../../utils/utils'
+import { setObjectColor } from '../../utils'
 
 const paramsMap: Param[] = [].sort((a, b) => a.index - b.index)
